Update margins logic
Discard margin handling in the printing driver, which used to add mental
load when trying to figure out the logic of how margins are implemented.
Having all margin handling in "userspace" of render engines (i.e. not in
the print driver abstraction level), will enable us to implement more
complicated logic in the future.