Refactor DymoLabeler classes
Fix a bug in which we had a constant for vertical margin, which is not true, since for every label width the margin is different.
Remove labeler state from DymoLabelerFunctions (it no longer initialized with current tape size).
Store default tape sizes in the labeler class, instead of in the GUI's code.